fix: use VALID_SESSION_INPUT in helpers, add restoreMocks to config
