+++ __tests__/fixtures/session.ts
@@ -1,0 +1,14 @@
+export const VALID_SESSION_INPUT = {
+  session_type: "regular",
+  title: "Week 1",
+  date: "2025-03-15",
+  time_start: "18:00",
+  time_end: "20:00",
+  location_name: "Gym",
+  location_address: "123 Main St",
+  location_maps_link: null,
+  player_cap: 12,
+  signup_open: "2025-03-14T09:00",
+  signup_close: "2025-03-15T17:00",
+  notes: null,
+};
