+++ __tests__/fixtures/user.ts
@@ -1,0 +1,3 @@
+import type { User } from "@supabase/supabase-js";
+
+export const MOCK_USER = { id: "user-123" } as User;
