+++ vitest.config.mts
@@ -9,6 +9,7 @@ export default defineConfig({
     },
   },
   test: {
+    restoreMocks: true,
     env: {
       NEXT_PUBLIC_SUPABASE_URL: "http://localhost:54321",
       NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY: "test-key",
